-Eye Glows are now a Character Option! Choose from None, Default, and Death Knight! Known Bug: Currently defaults to None on character load. -LWO Exporter:  -New M2 Exporter now serving as main LWO M2 Exporter.  -M2 Exporter now supports Luminous surfaces.  -Added Options to disable Camera & Bone exporting. (Not Yet Functional)  -(Dev Work) Doodads Sets as Layers exporting is 95% done. There's currently a bug that malforms some models. (The DwarvenBones01.mdx from altarofstorms, for example. Oddly, enough, it works perfectly when just exporting the DwarvenBones01.m2 model.)  -(Dev Work) Major reworking of various structures for better support.  -(Dev Work) Started work on new Scene-writing functions. -Dev Work:  -64-bit GLEW libs renamed to glew64.lib.  -Increased 64-bit compatibility.  -Added Matrix to Quaternion function.  -M2's animGeometry, animTextures and animBones boolans are now public.

+++ src/modelcontrol.cpp
@@ -225,7 +225,7 @@ void ModelControl::Update()
 	clbGeosets->Set(geosetItems, 0);
 
 	for (size_t i=0; i<model->geosets.size(); i++) {
-		clbGeosets->Check(i, model->showGeosets[i]);
+		clbGeosets->Check((unsigned int)i, model->showGeosets[i]);
 	}
 
 	bones->SetValue(model->showBones);
@@ -318,7 +318,7 @@ void ModelControl::OnList(wxCommandEvent &event)
 	int id = event.GetId();
 	if (id == ID_MODEL_GEOSETS) {
 		for (size_t i=0; i<model->geosets.size(); i++) {
-			model->showGeosets[i] = clbGeosets->IsChecked(i);
+			model->showGeosets[i] = clbGeosets->IsChecked((unsigned int)i);
 		}
 	}
 }
